Refactor SwishConfig for certificate-based authentication
Updated the SwishConfig class to use a client certificate filename and password instead of client ID and secret. Removed the ClientId and ClientSecret properties and added ClientCertificateFilename and ClientCertificatePassword.

Modified the AddSwishClient method in ServiceCollectionExtensions to accept an IConfiguration parameter, allowing retrieval of SwishConfig from configuration settings and adding it to the service collection as a singleton.

Adjusted the creation of the X509Certificate2 object to utilize the new properties from SwishConfig.